Remove chmod use and fix PikaCmd portability

+++ tools/testAllConfigs.sh
@@ -1,16 +1,15 @@
-#!/bin/bash
-
+#!/usr/bin/env bash
 set -e -o pipefail -u
-cd "${0%/*}"
+cd "$(dirname "$0")"
 
 tmp=$(mktemp /tmp/test.cppXXXX)
 echo 'int main() { return 0; }' >"$tmp"
 
 test_model() {
-    if ./BuildCpp.sh debug "$1" ../output/tmp_"$1" "$tmp" >/dev/null 2>&1; then
+    if bash ./BuildCpp.sh debug "$1" ../output/tmp_"$1" "$tmp" >/dev/null 2>&1; then
         rm -f ../output/tmp_"$1"
-        ./buildAndTest.sh debug "$1"
-        ./buildAndTest.sh release "$1"
+        bash ./buildAndTest.sh debug "$1"
+        bash ./buildAndTest.sh release "$1"
     else
         echo "Skipping $1 - compiler not available"
     fi
